Move fromCallback tests back to rx
